requirements.txt
pandas
Pillow
openpyxl
python-dotenv